sgtl5000: Rebase and add further sample rates.
The fs=x paramter is replaces by sample_rate=x, with sample_rate being
one of 8000, 10050, 12000, 16000, 22050, 24000, 32000, 44100 or 48000.

The additional argument mclk_mode tells the clock vs bit rate factor.
Suitable values are 0 or 3. With a value of 3, the internal clock
is generated by the PLL, and the argument mclk has to tell the
actual mclk frequency. That frequency can e.g. be generated by
PWM. The range is 8MHz - 27MHz. In the mode mclk_mode=0, the
mclk frequency is assumed to be sample_rate * 256.